test-requirements.txt
pytest==3.2.1
pytest-qt==2.1.2
pytest-cov==2.5.1
pytest-mock==1.6.2
pytest-pylint==0.7.1
python-dateutil==2.6.1
pytz==2017.2
keyring==10.4.0
mock==2.0.0